Fixed spelling / grammatical errors
Attempted to find and fix all spelling and grammatical errors. Also, Punctuation after a link should work fine and not become part of the link (except for invites for some reason, they turn it in to a normal link embed instead). And "Wi-Fi" is spelled with a hyphen from what I can find, but everyone (including ISPs) spells it wrong.